Implements Pina-Golada
This will be the foundation of PinaGolada. The FileStream struct will be
able to map files to their parsed ast struct as well as filter

Adds ast inspector

This commit adds an ast inspector that is capable of streaming ast objects based on a file stream.
The ast inspector is currently capable of reading/parsing from a file
stream as well as converting the stream into interfaces, which will be
useful for Pina-Golada

Adds the go file generator

This commit adds the go file generator which is capable of generating go
files based on method calls.

The interfaces are heavily based on the builder patterns to allow an
easy and elegant creation of go files

Introduces compressor framework

This commit introduces a compression framework. For now the only
compression format provided is .tar files, but we may add more in the
future

Implements the annotation parser

This commit implements the annotation parser on a CSV level. Other implementations may be possible, but at this point a proof of work is the goal and CSV works just fine

Rrefactors compressor

This commit refactors the compressor by implementing a virtual file
system the compressor can work with

Improves the virtual file system

This commit improves the virtual file system structure

Implements the generate command and builder

This commit implements the generate cobra command as well as the file
builder for a specific interface.

Adapts paths to multi os system

As paths had had trouble with different os implementations, this commit
fixes the trouble of / vs \ by parsing the path separator first and the
replacing it

Initialize Travis

Updates README

This commit updates the readme file to contain a small tutorial on how
to use Pina-Golada

Fixes relative path issue

Relative paths were not being evaluated, if not started with a ".".
As most of pina-golada will refer to relative paths, it should and now
is the default rather than the exception.

Updates comments

Reverts all the comments to a double dash on every line instead of
pretty multi block comments

Optimizes comments

Adds a slash after each // comment

Updates README

The readme file contained the wrong tool name on cli level. This commit fixes this issue