Introduces method to write directory to disk (#9)
* Introduces method to write directory to disk

Writes a virtual directory to the disk to improve assets storing on
disk. Checks if the directory is a root directory in which case it will
write everything in the directory to the path. If not the WriteToDisk
method will first create a sub-directory in the path, in which it stores
the contents of the passed virtual directory