Fixes builder directory tree
The builder used to have its own loader logic as it stores the folder to
load as a sub directory below the root diretory to keep the diretory
name present.

Now the builder does this prior to calling the default LoadFromDisk
method which fixes the lookup issues in the generated asset providers.