Implements integration test
As pina-golada did struggle a lot in the latests releases, this commit
adds an integration test to the project which compiles a test project
and tests for correct compilation and write to disk functionality.

Adds default permission to root directory

If someone writes the directory into another directory that does not yet
exist, this is the default permission that will be used, and currently
defaults to 0666.

Removed go 1.11 travis builds

As golang 1.12 has been established by now, this old version is simply
not needed for travis ci.